classname typo in readme
The full class name is GnugatMicroFrameworkBundle , not MicroFrameworkBundle, as the readme says. 

I'm presuming the lesser evil here is to change the readme to match the code, rather than vice versa?

PS - As a beginner to Symfony (pretty familiar with PHP), I'm really enjoying this bundle. Thanks for it! I look forward to making more contributions in the future.